[FIX] use a better regex for CMAKE_SYSTEM_PROCESSOR
According to CMake doc, CMAKE_SYSTEM_PROCESSOR is set to the
output of the command "uname -p" (in fact it is the output of "uname
-m"). This is only true for native compilation.

In the case of cross-compilation, the CMake code does not set anything,
it just expects the CMAKE_SYSTEM_* variables to be set in the
toolchain file.

Add a test matching the patern "^i.86$" and keep the old one "x86" for
compatibility.
By doing that, factorize this test using "^(i.86|x86(_64)?)$".

Thanks to Samuel Martin [1].

The issue has been reported [2] upstream an this patch will be merged
for the next release.

[1] http://lists.busybox.net/pipermail/buildroot/2014-November/112163.html
[2] http://sourceforge.net/p/openpowerlink/discussion/newbie/thread/3f13af65/#8f80/9f7c

Signed-off-by: Romain Naour <romain.naour@openwide.fr>