[FIX] install the stack libraries to "lib" subdirectory
Using '.' to install the stack libraries is not correct since
it will install them to /usr/.

ls /usr/liboplkmnapp-kernelintf.so

Fix this by using 'lib' instead of '.'

The problem with the openpowerlink build system is that it expect
the stack directory to be a "local staging" directory which
contains headers and libraries. So the libraries build by the
stack project are installed in stack/lib by default.

This is wrong to do that because it becomes difficult to package
openpowerlink for normal Linux systems.

Buildroot like others build system expect from a packaged software
to install libraries and headers build/provided in a staging
directory.

It seems that CMAKE_INSTALL_PREFIX is misused by the default value
"${CMAKE_CURRENT_SOURCE_DIR}/lib/${SYSTEM_NAME_DIR}/${SYSTEM_PROCESSOR_DIR}"
since it already contains the lib directory and therefore the
entire build system is messed up.

It seems that the openpowerlink's build system doesn't allow/support
CMAKE_INSTALL_PREFIX to be set externally
(which is the standard approach).

The issue has been reported [1] upstream and a build system
refactoring is needed to fix this properly.

[1] http://sourceforge.net/p/openpowerlink/discussion/newbie/thread/3f13af65/#8f80/a50a/4649/f07c

Signed-off-by: Romain Naour <romain.naour@openwide.fr>