[FIX] install the stack libraries to "lib" subdirectory
Using '.' to install the stack libraries is not correct since
it will install them to /usr/ when CMAKE_INSTALL_PREFIX is set
to "/usr/".

ls /usr/liboplkmnapp-kernelintf.so

Fix this by using 'lib' instead of '.'

The issue has been reported [1] upstream and a build system
refactoring is needed to fix this properly.

[1] http://sourceforge.net/p/openpowerlink/discussion/newbie/thread/3f13af65/#8f80/a50a/4649/f07c

Signed-off-by: Romain Naour <romain.naour@openwide.fr>